util to hash header
also added visp icon
