+++ src/qt/bitcoin.qrc
@@ -23,6 +23,7 @@
         <file alias="editpaste">res/icons/editpaste.png</file>
         <file alias="editcopy">res/icons/editcopy.png</file>
         <file alias="add">res/icons/add.png</file>
+		<file alias="visp">res/icons/visp.png</file>
         <file alias="bitcoin_testnet">res/icons/verge_testnet.png</file>
         <file alias="toolbar_testnet">res/icons/toolbar_verge_testnet.png</file>
         <file alias="edit">res/icons/edit.png</file>
